Dain patches for compilation and cleanup
git-svn-id: https://mclient-mume.googlecode.com/svn/trunk/mclient@252 962f5b74-7ec1-11dd-90a9-6306d3ffb32f

+++ build.sh
@@ -1,19 +1,17 @@
 #!/bin/bash
 declare -i JFLAG
-OSTYPE=`set | grep OSTYPE | cut -d = -f 1`
-if [ "$OSTYPE" == "linux-gnu" ]; then
-    PROCESSORS=`grep processor /proc/cpuinfo | wc -l`
+if [ -e /proc/cpuinfo ]; then
+    PROCESSORS=`grep '^processor\s*:' /proc/cpuinfo | wc -l`
     JFLAG=$PROCESSORS+1
 else
     JFLAG=2
 fi
 
-if [ -d build ]; then
-    rm -r build
-fi
+[ -d build ] && rm -r build
 
-mkdir build
-cd build
+mkdir -p build && cd build || exit 1
+
 cmake ../ -DCMAKE_BUILD_TYPE=Debug -DCMAKE_INSTALL_PREFIX=. \
--DMCLIENT_BIN_DIR=. -DMCLIENT_PLUGINS_DIR=plugins && make -j$JFLAG && make install
+-DMCLIENT_BIN_DIR=. -DMCLIENT_PLUGINS_DIR=plugins &&        \
+make -j$JFLAG && make install
 ln -sf ../config config
